Refactored the structure of the code - no code changes
Refactored the entire structure of the code for each question to be present under a specific topic.

This will help me organize the problems better which will have multiple solutions based on multiple topics.
