--- src/main/java/leetcode/q215/KthLargestElementInArray.java
@@ -1,6 +1,0 @@
-package leetcode.q215;
-public class KthLargestElementInArray {
-    public int findKthLargest(int[] nums, int k) {
-        return -1;
-    }
-}
