+++ src/main/java/leetcode/sorting_searching/q34/q162/PeakElement.java
@@ -1,0 +1,19 @@
+package leetcode.sorting_searching.q34.q162;
+public class PeakElement {
+    public int findPeakElement(int[] nums) {
+        int left = 0;
+        int right = nums.length - 1;
+
+        while (left < right) {
+            int mid = left + (right - left) / 2;
+
+            if (nums[mid] < nums[mid + 1]) {
+                left = mid + 1;
+            } else {
+                right = mid;
+            }
+        }
+
+        return left;
+    }
+}
